@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const api = axios.create({
-    baseURL: process.env.VUE_APP_ENDPOINT || 'http://zapia.com.br:1100',
+    baseURL: import.meta.env.VITE_VUE_APP_ENDPOINT || 'http://zapia.com.br:1100',
     headers: { 'Authorization': 'Bearer ' + (sessionStorage.TOKEN || '') }
 });
 
