@@ -1,5 +1,5 @@
 <template>
-    <span :inner-html.prop="lastMsg.body | emojify" v-if="lastMsg.isChat"></span>
+    <span v-html="msgBody" v-if="lastMsg.isChat"></span>
 
     <span v-else-if="lastMsg.isImage">
         <img src="@/assets/images/wpp-type-foto.svg">
@@ -40,7 +40,7 @@
 
     <span v-else-if="lastMsg.isVcard">
         <img src="@/assets/images/wpp-type-vcard.svg">
-        <span :inner-html.prop="lastMsg.vCard.fn[0].value | emojify"></span>
+        <span v-html="filters.emojify(lastMsg.vCard.fn[0].value)"></span>
     </span>
 
     <span v-else-if="lastMsg.isLocation">
@@ -64,6 +64,10 @@
 </template>
 
 <script>
+import filters from '@/filters';
+import { asyncComputed } from '@/AsyncComputed';
+import { useStore } from 'vuex';
+
 export default {
     name: 'MessageBody',
     props: {
@@ -71,6 +75,33 @@ export default {
             type: Object,
             required: true
         }
+    },
+    setup (props) {
+        const store = useStore();
+
+        const msgBody = asyncComputed(async () => {
+            let body = filters.emojify(filters.formatMsg(props.lastMsg.body));
+            if (props.lastMsg.mentionedJidList && props.lastMsg.mentionedJidList.length > 0) {
+                let promises = [];
+                let results = {};
+                for (let x = 0; x < props.lastMsg.mentionedJidList.length; x++) {
+                    promises.push(store.dispatch('findContactFromId', { id: props.lastMsg.mentionedJidList[x] }).then(contact => {
+                        results[props.lastMsg.mentionedJidList[x]] = contact;
+                    }));
+                }
+                await Promise.all(promises);
+                for (let x = 0; x < props.lastMsg.mentionedJidList.length; x++) {
+                    let contact = results[props.lastMsg.mentionedJidList[x]];
+                    let name = contact.formattedName || contact.verifiedName || contact.pushname;
+                    body = body.replace('@' + props.lastMsg.mentionedJidList[x].split('@')[0], `<span class='mention-symbol'>@</span><span dir="ltr">${name}</span>`);
+                }
+            }
+            return filters.emojify(body);
+        }, { default: props.lastMsg.body, lazy: true });
+
+        return {
+            msgBody
+        };
     },
     computed: {
         duration () {
@@ -88,7 +119,8 @@ export default {
 
             return `${m}:${s}`;
         }
-    }
+    },
+
 };
 </script>
 
