@@ -3,7 +3,7 @@
         :class="activeChat && activeChat.openChatInfo ? 'col-5 col-md-5 col-lg-5 col-xs-7 p-0' : 'col-8 col-md-8 col-lg-8 col-xs-9 p-0'"
         id="side-right">
         <ChatIntro v-if="!activeChat"/>
-        <Conversation :chat="activeChat" v-else/>
+        <Conversation v-else/>
     </div>
 </template>
 
