@@ -1,18 +1,19 @@
 <template>
     <div tabindex="0" id="conversation" @keydown="handleKeyDown">
-        <ConversationHeader :chat="chat"/>
+        <ConversationHeader/>
 
-        <MessagesList :chat="chat"/>
+        <MessagesList/>
 
-        <InputMessage :chat="chat"/>
+        <InputMessage/>
     </div>
 </template>
 
 <script>
+import { defineAsyncComponent } from 'vue';
 import ConversationHeader from './conversationHeader/ConversationHeader.vue';
-import InputMessage from './inputMessage/InputMessage';
-import MessagesList from './messagesList/MessagesList';
-import { mapState } from 'vuex';
+import MessagesList from './messagesList/MessagesList.vue';
+
+const InputMessage = defineAsyncComponent(() => import('./inputMessage/InputMessage.vue'));
 
 export default {
     name: 'Conversation',
@@ -21,18 +22,9 @@ export default {
         InputMessage,
         ConversationHeader
     },
-    props: {
-        chat: {
-            type: Object,
-            required: true
-        }
-    },
-    computed: {
-        ...mapState(['user'])
-    },
     methods: {
-        handleKeyDown (evt) {
-            this.$root.$emit('keyDown', evt);
+        handleKeyDown () {
+            //TODO: focus input on keydown
         }
     }
 };
