@@ -14,11 +14,11 @@
             </span>
 
             <span v-else-if=" isAdd ">
-                {{ isMe }} adicionou {{ formatedNumbers }}
+                {{ isMe }} adicionou {{ formattedNumbers }}
             </span>
 
             <span v-else-if=" isRemove ">
-                {{ isMe }} removeu {{ formatedNumbers }}
+                {{ isMe }} removeu {{ formattedNumbers }}
             </span>
 
             <span v-else-if=" isPromote ">
@@ -30,22 +30,19 @@
             </span>
 
             <span v-else-if=" isLeave ">
-                {{ formatedNumbers }} saiu
+                {{ formattedNumbers }} saiu
             </span>
 
             <span v-else-if=" isInvite ">
-                {{ formatedNumbers }} entrou usando o link de convite deste grupo
+                {{ formattedNumbers }} entrou usando o link de convite deste grupo
             </span>
         </div>
     </div>
 </template>
 
 <script>
-// e2e_notification identity
-// e2e_notification encrypt
-
-// gp2
-import { mapState, mapActions } from 'vuex';
+import { mapState, useStore } from 'vuex';
+import { asyncComputed } from '@/AsyncComputed';
 
 export default {
     name: 'MessageInfo',
@@ -55,19 +52,40 @@ export default {
             required: true
         }
     },
-    computed: {
-        ...mapState(['activeChat', 'self']),
+    setup (props) {
 
-        isMe () {
-            if (this.msg.senderObj && this.msg.senderObj.id === this.self.id) {
+        const store = useStore();
+
+        const formattedNumbers = asyncComputed(async () => {
+            if (props.msg.recipients && props.msg.recipients.length >= 1) {
+                let promises = props.msg.recipients.map((e) => {
+                    return store.dispatch('findFormattedNameFromId', { id: e });
+                });
+                let results = await Promise.all(promises);
+                return results.join(', ');
+            }
+            return '';
+        }, { default: props.msg.recipients ? props.msg.recipients.join(', ') : '', lazy: true });
+
+        const isMe = asyncComputed(async () => {
+            let senderObj = await props.msg.senderObj();
+            if (senderObj && senderObj.id === store.state.self.id) {
                 return 'Você';
-            } else if (this.msg.senderObj && this.msg.senderObj.formattedName) {
-                return this.msg.senderObj.formattedName;
-            } else if (this.msg.senderObj) {
-                return '+' + this.msg.senderObj.id.replace('@c.us', '');
+            } else if (senderObj && senderObj.formattedName) {
+                return senderObj.formattedName;
+            } else if (senderObj) {
+                return '+' + senderObj.id.replace('@c.us', '');
             }
             return 'Você';
-        },
+        }, { default: '', lazy: true });
+
+        return {
+            formattedNumbers,
+            isMe
+        };
+    },
+    computed: {
+        ...mapState(['activeChat', 'self']),
 
         isEncrypt () {
             return this.msg.type === 'e2e_notification' && this.msg.subtype === 'encrypt';
@@ -95,29 +113,6 @@ export default {
         },
         isInvite () {
             return this.msg.type === 'gp2' && this.msg.subtype === 'invite';
-        }
-    },
-    methods: {
-        ...mapActions(['findFormattedNameFromId'])
-    },
-    asyncComputed: {
-        formatedNumbers: {
-            async get () {
-                if (this.msg.recipients && this.msg.recipients.length >= 1) {
-                    let promises = this.msg.recipients.map((e) => {
-                        return this.findFormattedNameFromId({ id: e });
-                    });
-                    let results = await Promise.all(promises);
-                    return results.join(', ');
-                }
-                return '';
-            },
-            default () {
-                if (this.msg.recipients) {
-                    return this.msg.recipients.join(', ');
-                }
-                return '';
-            }
         }
     }
 };
