@@ -1,5 +1,9 @@
 <template>
-    <div class="message-video" v-b-visible.once="onVisible">
+    <div class="message-video" v-observe-visibility="{
+         throttle: 300,
+         callback: onVisible,
+         once: true
+    }">
         <div @click="handleClick" class="video-container">
             <div class="box-preview blur">
                 <img :src=" 'data:image/jpeg;base64,'+ msg.body" alt="body">
@@ -21,7 +25,7 @@
 import { mapActions, mapMutations } from 'vuex';
 import LoadingMedia from '../loadingMedia/LoadingMedia.vue';
 import MessageTime from '../messageTime/MessageTime.vue';
-import PlayMedia from '../playMedia/PlayMedia';
+import PlayMedia from '../playMedia/PlayMedia.vue';
 
 export default {
     name: 'MessageVideo',
@@ -56,7 +60,8 @@ export default {
                     await Promise.all(promises);
                     for (let x = 0; x < this.msg.mentionedJidList.length; x++) {
                         let chat = results[this.msg.mentionedJidList[x]];
-                        let name = chat.contact.formattedName || chat.contact.verifiedName || chat.contact.pushname;
+                        let contact = await chat.contact();
+                        let name = contact.formattedName || contact.verifiedName || contact.pushname;
                         caption = caption.replace('@' + this.msg.mentionedJidList[x].split('@')[0], `<span class='mention-symbol'>@</span><span class='btn-link' dir="ltr">${name}</span>`);
                     }
                 }
@@ -149,7 +154,7 @@ img {
     bottom: 3px;
 }
 
-.box-caption >>> .mention-symbol {
+.box-caption ::v-deep(.mention-symbol) {
     color: rgba(0, 0, 0, 0.25);
 }
 </style>
